+++ R/compute_minP.R
@@ -10,6 +10,10 @@
 compute_minP <- function(distribution, alternative) {
   distribution_rank <- apply(distribution,2,function(col){compute_all_pvalue(col,alternative = alternative)})
   minp <- apply(distribution_rank,1,min)
+
+  length_unique_minp <- length(unique(minp))
+  if(length_unique_minp<200)warnings(paste0("The adjusted distribution contains only ",length_unique_minp," unique values. Try to increase the number of permutation."))
+
   pvalue <- distribution_rank[1,]
   p_corrected <- sapply(pvalue,function(pi)compute_pvalue(distribution = minp,
                                                           stat = pi,alternative = "less"))