+++ NEWS.md
@@ -8,7 +8,8 @@
  * rcpp optimization: get_cluster(), compute_tfce()
  * add compute_clusterdepth()
  * add alternative to compute_tfce()
- 
+ * add warnings to minp and troendle
+
 
 # permuco 1.1.0 (cran)